requirements_sparq.txt
# SGM Agent Requirements

# Core dependencies
numpy>=1.20.0
pyyaml>=5.4.0

# Optional: Fast vector search (recommended for production)
# Uncomment one of the following:
# faiss-cpu>=1.7.0
# faiss-gpu>=1.7.0

# Optional: Visualization and analysis
matplotlib>=3.3.0
pandas>=1.2.0

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0

# For AgentGym integration (install separately)
# Follow AgentGym installation instructions:
# https://github.com/WooooDyy/AgentGym
